Fix compiler warnings and errors
- Fix unused parameter warnings in logger.h by commenting out parameter names
- Fix sign comparison warning in thread.h by casting elapsed to unsigned long
- Fix int-to-pointer cast warnings in canmanager.cpp using uintptr_t
- Remove deprecated Qt::AA_EnableHighDpiScaling in main.cpp (always enabled in Qt 6)
- Replace deprecated QVariant::type() with typeId() in amjsonsignal.cpp
- Fix QString/std::string conversion in graphicitemsenummap.cpp

+++ canquick/canmanager.cpp
@@ -540,15 +540,15 @@ void CanManager::read_frame(void)
     else
     {
 #if 1
-        coreDebug() << "can interface:" << (void*) (quint32) frame.can_id << ":" <<
-                   (void*) (quint32) frame.data[0] <<
-                   (void*) (quint32) frame.data[1] <<
-                   (void*) (quint32) frame.data[2] <<
-                   (void*) (quint32) frame.data[3] <<
-                   (void*) (quint32) frame.data[4] <<
-                   (void*) (quint32) frame.data[5] <<
-                   (void*) (quint32) frame.data[6] <<
-                   (void*) (quint32) frame.data[7] <<
+        coreDebug() << "can interface:" << (void*) static_cast<uintptr_t>(frame.can_id) << ":" <<
+                   (void*) static_cast<uintptr_t>(frame.data[0]) <<
+                   (void*) static_cast<uintptr_t>(frame.data[1]) <<
+                   (void*) static_cast<uintptr_t>(frame.data[2]) <<
+                   (void*) static_cast<uintptr_t>(frame.data[3]) <<
+                   (void*) static_cast<uintptr_t>(frame.data[4]) <<
+                   (void*) static_cast<uintptr_t>(frame.data[5]) <<
+                   (void*) static_cast<uintptr_t>(frame.data[6]) <<
+                   (void*) static_cast<uintptr_t>(frame.data[7]) <<
                    "ts:" << QDateTime::currentMSecsSinceEpoch();
 #endif
